Move fixture to separate directory
